requirements.txt
selenium>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.1
fake-useragent>=1.5.1
supabase>=2.5.0
undetected-chromedriver>=3.5.4
flask>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
gunicorn>=21.0.0
